Refine store implementation
